Read info from several more X-Com datafiles. Respect the terrain type of X-Com loads. Moved all world texture definitions to hex. Cleaned up some of the texturing code.

+++ world.h
@@ -60,8 +60,8 @@ protected:
 
   vector<SimpleModel *> models;
   vector<SimpleTexture *> textures;
-  map<Uint8, unsigned int> modmap;
-  map<Uint8, unsigned int> texmap;
+  map<int, unsigned int> modmap;
+  map<int, unsigned int> texmap;
   int smoke, fire;
   };
